requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
Werkzeug==2.2.2
pika==1.3.1
sqlalchemy==2.0.9